Move tests for large files to a new manual tests project.
